split jblas file. add compute type fp32.
