+++ intel_extension_for_transformers/llm/runtime/graph/core/jblas_common.cpp
@@ -1,0 +1,23 @@
+#include "jblas_common.hpp"
+using namespace jblas;
+
+void jblas_init() {
+  GetCPUDevice();
+  if (_cd->AMX_BF16() || _cd->AMX_INT8()) {
+    utils::request_perm_xtile_data();
+  }
+  _cd->print();
+}
+
+void jblas_timer(bool _init) {
+  static utils::timer<utils::microseconds> tr;
+  if (_init)
+    tr.start();
+  else
+    printf("time :%f us\n", tr.stop());
+}
+
+int jblas_set_threads(int _nth) {
+  jblas::utils::parallel::CpuDevice::getInstance()->setThreads(_nth);
+  return jblas::utils::parallel::CpuDevice::getInstance()->getThreads();
+}
