+++ intel_extension_for_transformers/llm/runtime/graph/core/layers/ip_fusion_qkv.cpp
@@ -1,0 +1,233 @@
+#include "jblas_common.hpp"
+
+using namespace jblas;
+using namespace ne_jblas;
+
+namespace {
+namespace transformer {
+namespace avx512_vnni {
+static JBLAS_ISA constexpr DefaultISA = JblasAVX512_VNNI;
+using QKVGemmDynamicS4Fp32KBlock = jblas::wrapper::transformer::QKVGemmInterfaceKBlockPackWeight<
+    jblas::wrapper::gemm_kblock::GemmSLauncherKBlockPackWeight<
+        DefaultISA, jblas::gemm::kblock::GemmCore_Row_NN_3x48_AVX512_VNNI_KBLOCK,
+        jblas::prologue::gemm::ActivationF32U8KBlockQuantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS4ClipScaleFp32,
+        jblas::epilogue::gemm::AccumulatorWriteBackFp32>,
+    jblas::utils::parallel::Parallel2DGemmKBlockFixed>;
+using QKVGemmDynamicS8Fp32KBlock = jblas::wrapper::transformer::QKVGemmInterfaceKBlockPackWeight<
+    jblas::wrapper::gemm_kblock::GemmSLauncherKBlockPackWeight<
+        DefaultISA, jblas::gemm::kblock::GemmCore_Row_NN_3x48_AVX512_VNNI_KBLOCK,
+        jblas::prologue::gemm::ActivationF32U8KBlockQuantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32, jblas::epilogue::gemm::AccumulatorWriteBackFp32>,
+    jblas::utils::parallel::Parallel2DGemmKBlockFixed>;
+using QKVGemmDynamicS8Fp32PerN = jblas::wrapper::transformer::QKVGemmInterfacePackWeightParallelAB<
+    jblas::wrapper::gemm_pack_weight::GemmLauncherPackWeight<
+        DefaultISA, jblas::gemm::GemmCore_Row_NN_8x48_AVX512_VNNI, jblas::prologue::gemm::ActivationFp32AsymU8Quantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32PerChannelN,
+        jblas::epilogue::gemm::ZpDequantInt32ToFp32>,
+    jblas::utils::parallel::Parallel2DGemm>;
+}  // namespace avx512_vnni
+namespace amx_int8 {
+static JBLAS_ISA constexpr DefaultISA = JblasAMX_INT8;
+using QKVGemmDynamicS4Fp32KBlock = jblas::wrapper::transformer::QKVGemmInterfaceKBlockPackWeight<
+    jblas::wrapper::gemm_kblock::GemmSLauncherKBlockPackWeight<
+        DefaultISA, jblas::gemm::kblock::GemmCore_Row_NN_16x48_AMX_INT8_KBLOCK,
+        jblas::prologue::gemm::ActivationF32S8KBlockQuantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS4ClipScaleFp32,
+        jblas::epilogue::gemm::AccumulatorWriteBackFp32>,
+    jblas::utils::parallel::Parallel2DGemmKBlockFixed>;
+using QKVGemmDynamicS8Fp32KBlock = jblas::wrapper::transformer::QKVGemmInterfaceKBlockPackWeight<
+    jblas::wrapper::gemm_kblock::GemmSLauncherKBlockPackWeight<
+        DefaultISA, jblas::gemm::kblock::GemmCore_Row_NN_16x48_AMX_INT8_KBLOCK,
+        jblas::prologue::gemm::ActivationF32S8KBlockQuantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32, jblas::epilogue::gemm::AccumulatorWriteBackFp32>,
+    jblas::utils::parallel::Parallel2DGemmKBlockFixed>;
+using QKVGemmDynamicS8Fp32PerN = jblas::wrapper::transformer::QKVGemmInterfacePackWeightParallelAB<
+    jblas::wrapper::gemm_pack_weight::GemmLauncherPackWeight<
+        DefaultISA, jblas::gemm::GemmCore_Row_NN_16x48_AMX_S8S8, jblas::prologue::gemm::ActivationFp32SymS8Quantize,
+        jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32PerChannelN,
+        jblas::epilogue::gemm::DequantInt32ToFp32>,
+    jblas::utils::parallel::Parallel2DGemm>;
+}  // namespace amx_int8
+}  // namespace transformer
+}  // namespace
+
+bool jblas_fusion_QKV_f32f32_support(void* wqptr, void* wkptr, void* wvptr, int _m, int _n, int _k) {
+  auto wqtmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wqptr, 0);
+  auto wktmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wkptr, 0);
+  auto wvtmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wvptr, 0);
+  bool support = false;
+  if (wqtmp != nullptr && wktmp != nullptr && wvtmp != nullptr) {
+    prologue::PackedWeight* tmps[3] = {wqtmp, wktmp, wvtmp};
+    auto sameKernel = samePackedWeight(tmps, 3);
+    if (sameKernel) {
+      if (wqtmp->mType == int(WeightCompType::WeightS4ClipScaleFp32) ||
+          wqtmp->mType == int(WeightCompType::WeightS8ScaleFp32)) {
+        constexpr size_t EleNum = sizeof(GcCompInt8KBlockSet) / sizeof(GcCompInt8KBlockSet[0]);
+        support = contains(wqtmp->mCoreType, GcCompInt8KBlockSet, EleNum);
+        support &= hasISA(GcCompInt8KBlockSet, EleNum);
+      } else if (wqtmp->mType == int(WeightCompType::WeightS8ScaleFp32PerChannelN)) {
+        constexpr size_t EleNum = sizeof(GcCompInt8Set) / sizeof(GcCompInt8Set[0]);
+        support = contains(wqtmp->mCoreType, GcCompInt8Set, EleNum);
+        support &= hasISA(GcCompInt8Set, EleNum);
+      }
+    }
+  }
+  safe_delete(wqtmp);
+  safe_delete(wktmp);
+  safe_delete(wvtmp);
+  return support;
+}
+
+JBLAS_CODE jblas_QKVs4fp32_f32f32_forward(float* activation, SS4Fp32* wqptr, SS4Fp32* wkptr, SS4Fp32* wvptr,
+                                          float* output, int _m, int _n, int _k, int lda, int ldo) {
+  GetCPUDevice();
+  auto ret = JblasRuntimeError;
+  if (wqptr->mCoreType == GcCompInt8KBlock::TYPE) {
+    if (_cd->AMX_INT8() && wqptr->mBlockSize % 128 == 0) {
+      using GemmKernel = jblas::wrapper::transformer_default::weight_comp::amx_int8::QKVGemmDynamicS4Fp32KBlock;
+      static GemmKernel kernel;
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo},
+          {output + _m * _n, ldo},
+          {output + 2 * _m * _n, ldo},
+      };
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, wqptr->mBlockSize, NULL);
+      ret = kernel.compute({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    } else if (_cd->AVX512_VNNI()) {
+      using GemmKernel = jblas::wrapper::transformer_default::weight_comp::avx512_vnni::QKVGemmDynamicS4Fp32KBlock;
+      static GemmKernel kernel;
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo},
+          {output + _m * _n, ldo},
+          {output + 2 * _m * _n, ldo},
+      };
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, wqptr->mBlockSize, NULL);
+      ret = kernel.compute({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    }
+  }
+  return ret;
+}
+
+JBLAS_CODE jblas_QKVs8fp32_f32f32_forward(float* activation, SS8Fp32* wqptr, SS8Fp32* wkptr, SS8Fp32* wvptr,
+                                          float* output, int _m, int _n, int _k, int lda, int ldo) {
+  GetCPUDevice();
+  auto ret = JblasRuntimeError;
+  if (wqptr->mCoreType == GcCompInt8KBlock::TYPE) {
+    if (_cd->AMX_INT8() && wqptr->mBlockSize % 128 == 0) {
+      using GemmKernel = transformer::amx_int8::QKVGemmDynamicS8Fp32KBlock;
+      static GemmKernel kernel;
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo},
+          {output + _m * _n, ldo},
+          {output + 2 * _m * _n, ldo},
+      };
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, wqptr->mBlockSize, NULL);
+      ret = kernel.compute({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    } else if (_cd->AVX512_VNNI()) {
+      using GemmKernel = transformer::avx512_vnni::QKVGemmDynamicS8Fp32KBlock;
+      static GemmKernel kernel;
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo},
+          {output + _m * _n, ldo},
+          {output + 2 * _m * _n, ldo},
+      };
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, wqptr->mBlockSize, NULL);
+      ret = kernel.compute({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    }
+  }
+  return ret;
+}
+
+JBLAS_CODE jblas_QKVs8fp32pern_f32f32_forward(float* activation, SS8Fp32PerN* wqptr, SS8Fp32PerN* wkptr,
+                                              SS8Fp32PerN* wvptr, float* output, int _m, int _n, int _k, int lda,
+                                              int ldo) {
+  GetCPUDevice();
+  auto ret = JblasRuntimeError;
+  if (wqptr->mCoreType == GcCompInt8::TYPE) {
+    if (_cd->AMX_INT8()) {
+      using GemmKernel = transformer::amx_int8::QKVGemmDynamicS8Fp32PerN;
+      static GemmKernel kernel;
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, NULL);
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo, quanA->mSPtr, quanA->lds, wqptr->mSPtr},
+          {output + _m * _n, ldo, quanA->mSPtr, quanA->lds, wkptr->mSPtr},
+          {output + 2 * _m * _n, ldo, quanA->mSPtr, quanA->lds, wvptr->mSPtr},
+      };
+      ret = kernel.compute<true, false>({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    } else if (_cd->AVX512_VNNI()) {
+      using GemmKernel = transformer::avx512_vnni::QKVGemmDynamicS8Fp32PerN;
+      static GemmKernel kernel;
+      auto quanA = kernel.getActivationPtr()->createStorage(_m, _k, NULL);
+      GemmKernel::WeightType::Param wparams[3]{
+          wqptr,
+          wkptr,
+          wvptr,
+      };
+      GemmKernel::CParam oparams[3]{
+          {output, ldo, quanA->mZPtr, quanA->mSPtr, quanA->lds, wqptr->mSPtr},
+          {output + _m * _n, ldo, quanA->mZPtr, quanA->mSPtr, quanA->lds, wkptr->mSPtr},
+          {output + 2 * _m * _n, ldo, quanA->mZPtr, quanA->mSPtr, quanA->lds, wvptr->mSPtr},
+      };
+      ret = kernel.compute<true, false>({_m, _n, _k, 3, activation, lda, quanA, wparams, oparams, NULL});
+      delete quanA;
+    }
+  }
+  return ret;
+}
+
+// f32f32: activation & output dtype
+void jblas_fusion_QKV_f32f32_forward(float* activation, void* wqptr, void* wkptr, void* wvptr, float* output, int _m,
+                                     int _n, int _k, int lda, int ldo) {
+  GetCPUDevice();
+  auto ret = JblasRuntimeError;
+  auto wqtmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wqptr, 0);
+  auto wktmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wkptr, 0);
+  auto wvtmp = prologue::weight_comp::gemm_kblcok::PackedWeightParser::deserialBuffer(wvptr, 0);
+  // must check support before forward, there is no need to check support twice.
+  if (wqtmp->mType == int(WeightCompType::WeightS4ClipScaleFp32)) {
+    ret = jblas_QKVs4fp32_f32f32_forward(activation, dynamic_cast<SS4Fp32*>(wqtmp), dynamic_cast<SS4Fp32*>(wktmp),
+                                         dynamic_cast<SS4Fp32*>(wvtmp), output, _m, _n, _k, lda, ldo);
+  } else if (wqtmp->mType == int(WeightCompType::WeightS8ScaleFp32)) {
+    ret = jblas_QKVs8fp32_f32f32_forward(activation, dynamic_cast<SS8Fp32*>(wqtmp), dynamic_cast<SS8Fp32*>(wktmp),
+                                         dynamic_cast<SS8Fp32*>(wvtmp), output, _m, _n, _k, lda, ldo);
+  } else if (wqtmp->mType == int(WeightCompType::WeightS8ScaleFp32PerChannelN)) {
+    ret = jblas_QKVs8fp32pern_f32f32_forward(activation, dynamic_cast<SS8Fp32PerN*>(wqtmp),
+                                             dynamic_cast<SS8Fp32PerN*>(wktmp), dynamic_cast<SS8Fp32PerN*>(wvtmp),
+                                             output, _m, _n, _k, lda, ldo);
+  }
+  assert(ret == JblasSuccess);
+  safe_delete(wqtmp);
+  safe_delete(wktmp);
+  safe_delete(wvtmp);
+}
