+++ intel_extension_for_transformers/llm/runtime/graph/models/gptj/gptj.cpp
@@ -32,7 +32,7 @@
 #include "core/data_types.h"
 #include "core/ne.h"
 #include "core/ne_layers.h"
-#include "core/layers/inner_product.h"
+#include "core/ne_jblas.h"
 #include "core/layers/mha_dense.h"
 #include "models/model_utils/model_config.h"
 #include "models/model_utils/model_utils.h"
