+++ intel_extension_for_transformers/llm/runtime/graph/models/model_utils/model_utils.cpp
@@ -42,6 +42,7 @@
 
 #include "core/ne_layers.h"
 #include "application/common.h"
+#include "core/jblas_common.hpp"
 #include "jblas/jblas/jit_blas_weight_compression.h"
 #include "models/model_utils/model_files.h"
 #include "models/model_utils/model_utils.h"
@@ -769,23 +770,10 @@ quant_params_internal quant_params_to_internal(const quant_params& params) {
   return quant_params_internal{parse_bits(params.bits), parse_alg(params.alg), params.block_size,
                                parse_scale_dtype(params.scale_dtype), parse_compute_type(params.compute_type)};
 }
-template <class T, JBLAS_ISA ISA>
-using WeiS4Fp32 = jblas::prologue::weight_comp::gemm_kblcok::WeightS4ClipScaleFp32<T, ISA>;
-
-template <class T, JBLAS_ISA ISA>
-using WeiS8Fp32 = jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32<T, ISA>;
-
-template <class T, JBLAS_ISA ISA>
-using WeiS8PerNFp32 = jblas::prologue::weight_comp::gemm_kblcok::WeightS8ScaleFp32PerChannelN<T, ISA>;
-
-using CompFp32 = jblas::gemm::GemmCore_Row_NN_8x48_AVX512F;
-using CompInt8 = jblas::gemm::kblock::GemmCore_Row_NN_3x48_AVX512_VNNI_KBLOCK;
-using CompBf16 = jblas::gemm::GemmCore_Row_NN_16x64_AMX_BF16;
-using CompFp16 = jblas::gemm::GemmCore_Row_NN_8x64_AVX512_FP16;
-using CompPureInt8 = jblas::gemm::GemmCore_Row_NN_8x48_AVX512_VNNI;
 
 size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_internal params, int nthread, int n, int k) {
   using CompType = jblas::prologue::weight_comp::gemm_kblcok::WeightCompType;
+  using namespace ne_jblas;
   auto cd = jblas::utils::parallel::CpuDevice::getInstance();
   jblas::prologue::PackedWeight* packedw = NULL;
 
@@ -793,8 +781,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
   if (params.bits == quant_bits::q4) {
     if (params.scale_dtype == quant_sdtype::fp32) {
       if (params.compute_type == quant_comp::int8) {
-        using Kernel = WeiS4Fp32<CompInt8, JblasAVX512F>;
-        using KernelRef = WeiS4Fp32<CompInt8, JblasNoSIMD>;
+        using Kernel = WeiS4ClipFp32<GcCompInt8KBlock, JblasAVX512F>;
+        using KernelRef = WeiS4ClipFp32<GcCompInt8KBlock, JblasNoSIMD>;
         static Kernel kernel;
         static Kernel kernelref;
         packedw = kernel.createStorage(n, k, params.block_size);
@@ -804,8 +792,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
           kernelref.packTransposeWeight(n, k, f32ptr, k, packedw);
         }
       } else if (params.compute_type == quant_comp::fp32) {
-        using Kernel = WeiS4Fp32<CompFp32, JblasAVX512_FP16>;
-        using KernelRef = WeiS4Fp32<CompFp32, JblasNoSIMD>;
+        using Kernel = WeiS4ClipFp32<GcCompFp32, JblasAVX512_FP16>;
+        using KernelRef = WeiS4ClipFp32<GcCompFp32, JblasNoSIMD>;
         static Kernel kernel;
         static Kernel kernelref;
         packedw = kernel.createStorage(n, k, params.block_size);
@@ -815,8 +803,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
           kernelref.packTransposeWeight(n, k, f32ptr, k, packedw);
         }
       } else if (params.compute_type == quant_comp::bf16) {
-        using Kernel = WeiS4Fp32<CompBf16, JblasAMX_BF16>;
-        using KernelRef = WeiS4Fp32<CompBf16, JblasNoSIMD>;
+        using Kernel = WeiS4ClipFp32<GcCompBf16, JblasAMX_BF16>;
+        using KernelRef = WeiS4ClipFp32<GcCompBf16, JblasNoSIMD>;
         static Kernel kernel;
         static Kernel kernelref;
         packedw = kernel.createStorage(n, k, params.block_size);
@@ -833,8 +821,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
     if (params.scale_dtype == quant_sdtype::fp32) {
       if (params.compute_type == quant_comp::int8) {
         if (params.block_size == -1) {
-          using Kernel = WeiS8PerNFp32<CompPureInt8, JblasAVX512F>;
-          using KernelRef = WeiS8PerNFp32<CompPureInt8, JblasNoSIMD>;
+          using Kernel = WeiS8Fp32PerN<GcCompInt8, JblasAVX512F>;
+          using KernelRef = WeiS8Fp32PerN<GcCompInt8, JblasNoSIMD>;
           static Kernel kernel;
           static Kernel kernelref;
           packedw = kernel.createStorage(n, k);
@@ -844,8 +832,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
             kernelref.packTransposeWeight(n, k, f32ptr, k, packedw);
           }
         } else {
-          using Kernel = WeiS8Fp32<CompInt8, JblasAVX512F>;
-          using KernelRef = WeiS8Fp32<CompInt8, JblasNoSIMD>;
+          using Kernel = WeiS8Fp32<GcCompInt8, JblasAVX512F>;
+          using KernelRef = WeiS8Fp32<GcCompInt8, JblasNoSIMD>;
           static Kernel kernel;
           static Kernel kernelref;
           packedw = kernel.createStorage(n, k, params.block_size);
@@ -856,8 +844,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
           }
         }
       } else if (params.compute_type == quant_comp::fp32) {
-        using Kernel = WeiS8Fp32<CompFp32, JblasAVX512_FP16>;
-        using KernelRef = WeiS8Fp32<CompFp32, JblasNoSIMD>;
+        using Kernel = WeiS8Fp32<GcCompFp32, JblasAVX512_FP16>;
+        using KernelRef = WeiS8Fp32<GcCompFp32, JblasNoSIMD>;
         static Kernel kernel;
         static Kernel kernelref;
         packedw = kernel.createStorage(n, k, params.block_size);
@@ -867,8 +855,8 @@ size_t jblas_quantize(const float* f32ptr, void* dstpr, const quant_params_inter
           kernelref.packTransposeWeight(n, k, f32ptr, k, packedw);
         }
       } else if (params.compute_type == quant_comp::bf16) {
-        using Kernel = WeiS8Fp32<CompBf16, JblasAMX_BF16>;
-        using KernelRef = WeiS8Fp32<CompBf16, JblasNoSIMD>;
+        using Kernel = WeiS8Fp32<GcCompBf16, JblasAMX_BF16>;
+        using KernelRef = WeiS8Fp32<GcCompBf16, JblasNoSIMD>;
         static Kernel kernel;
         static Kernel kernelref;
         packedw = kernel.createStorage(n, k, params.block_size);
